rename core.definitions.event_metadata to core.definitions.metadata
