+++ python_modules/dagster/dagster/core/definitions/__init__.py
@@ -30,7 +30,7 @@ from .dependency import (
     SolidInvocation,
     SolidOutputHandle,
 )
-from .event_metadata import (
+from .metadata import (
     EventMetadata,
     EventMetadataEntry,
     FloatMetadataEntryData,
