+++ python_modules/dagster/dagster/core/definitions/input.py
@@ -2,7 +2,7 @@ from collections import namedtuple
 from typing import NamedTuple, Optional, Set
 
 from dagster import check
-from dagster.core.definitions.event_metadata import EventMetadataEntry, parse_metadata
+from dagster.core.definitions.metadata import EventMetadataEntry, parse_metadata
 from dagster.core.definitions.events import AssetKey
 from dagster.core.errors import DagsterError, DagsterInvalidDefinitionError
 from dagster.core.types.dagster_type import (
