+++ python_modules/dagster/dagster/core/definitions/output.py
@@ -14,7 +14,7 @@ from typing import (
 )
 
 from dagster import check
-from dagster.core.definitions.event_metadata import EventMetadataEntry, parse_metadata
+from dagster.core.definitions.metadata import EventMetadataEntry, parse_metadata
 from dagster.core.definitions.events import AssetKey
 from dagster.core.errors import DagsterError, DagsterInvalidDefinitionError
 from dagster.core.types.dagster_type import DagsterType, resolve_dagster_type
