+++ python_modules/dagster/dagster/core/storage/fs_io_manager.py
@@ -4,7 +4,7 @@ import pickle
 from dagster import check
 from dagster.config import Field
 from dagster.config.source import StringSource
-from dagster.core.definitions.event_metadata import EventMetadataEntry
+from dagster.core.definitions.metadata import EventMetadataEntry
 from dagster.core.definitions.events import AssetKey, AssetMaterialization
 from dagster.core.errors import DagsterInvariantViolationError
 from dagster.core.execution.context.input import InputContext
