+++ python_modules/libraries/dagstermill/dagstermill/io_managers.py
@@ -4,7 +4,7 @@ from typing import Any, List, Optional
 
 from dagster import check
 from dagster.config.field import Field
-from dagster.core.definitions.event_metadata import EventMetadataEntry
+from dagster.core.definitions.metadata import EventMetadataEntry
 from dagster.core.definitions.events import AssetKey
 from dagster.core.execution.context.input import InputContext
 from dagster.core.execution.context.output import OutputContext
